Add files via upload
There were dozens of revisions and clarifications but NO corrections for the first time in a new version of a toolbox. The big addition is the CASE PATTERN MATCHING that became available in 3.10.  That said, this version has been drastically rearranged to be more logically grouped and easier to intuitively navigate once you have been through it a couple of times.  Enjoy!